build: avoid recovery patch if device setup can't do it correctly
In the case of Amazon OMAP4's there is a 2nd-bootloader which is inserted at
a fixed location in both boot.img and recovery.img.  If this 2nd-bootloader
is not present, it puts the device in a non-booting state requiring the user
to recover the device using fastboot tools.

Also, due to the extra binary information in the image, imgdiff hangs during
build.

For devices like this, allow bypass of the recovery patch generatation.

NOTE: In previous versions of CM the BOARD_CUSTOM_BOOTIMG_MK flag accomplished
this, but was never forward ported to cm-12.0 and on.  Use a new board flag
here for clarity as most custom bootimg makefiles can still process a patch
for recovery.
https://github.com/CyanogenMod/android_build/blob/cm-11.0/core/Makefile#L935

Change-Id: I9284ee227358224e3add9862db5c1fbd0f5bd226
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>